Appease the linter and move the GetSumOfHistogramSampleCount to ha_tracker_test.go
