+++ pkg/util/metrics_helper.go
@@ -10,38 +10,22 @@ import (
 	"github.com/prometheus/prometheus/model/labels"
 )
 
-// FromLabelPairsToLabels converts dto.LabelPair into labels.Labels.
-func FromLabelPairsToLabels(pairs []*dto.LabelPair) labels.Labels {
-	builder := labels.NewBuilder(nil)
-	for _, pair := range pairs {
-		builder.Set(pair.GetName(), pair.GetValue())
-	}
-	return builder.Labels(nil)
-}
-
-// GetSumOfHistogramSampleCount returns the sum of samples count of histograms matching the provided metric name
-// and optional label matchers. Returns 0 if no metric matches.
-func GetSumOfHistogramSampleCount(families []*dto.MetricFamily, metricName string, matchers labels.Selector) uint64 {
-	sum := uint64(0)
-
-	for _, metric := range families {
-		if metric.GetName() != metricName {
-			continue
-		}
-
-		if metric.GetType() != dto.MetricType_HISTOGRAM {
-			continue
-		}
-
-		for _, series := range metric.GetMetric() {
-			if !matchers.Matches(FromLabelPairsToLabels(series.GetLabel())) {
+func MatchesSelectors(m *dto.Metric, selectors labels.Labels) bool {
+	for _, l := range selectors {
+		found := false
+		for _, lp := range m.GetLabel() {
+			if l.Name != lp.GetName() || l.Value != lp.GetValue() {
 				continue
 			}
 
-			histogram := series.GetHistogram()
-			sum += histogram.GetSampleCount()
+			found = true
+			break
+		}
+
+		if !found {
+			return false
 		}
 	}
 
-	return sum
+	return true
 }
